Fix a strange director? issue
For some reason, in some situations, the handler callbacks are
being called through a specific director code path that adds
a function to the end of the argument list. This workaround
doesn't break anything, but is ugly and out of place.

The director code that adds the function is director/lib/director/router.js:407